edit

+++ flowmanage.h
@@ -7,9 +7,9 @@ class flowmanage
 public:
     uint32_t ip_src;
     uint32_t ip_dst;
-    uint8_t sport;
-    uint8_t dport;
-    flowmanage(const uint32_t &ip_src, const uint32_t &ip_dst, const uint8_t &sport, const uint8_t &dport)
+    uint16_t sport;
+    uint16_t dport;
+    flowmanage(const uint32_t &ip_src, const uint32_t &ip_dst, const uint16_t &sport, const uint16_t &dport)
     {
         this->ip_src = ip_src;
         this->ip_dst = ip_dst;
@@ -17,13 +17,14 @@ public:
         this->dport = dport;
     }
 
-    void init(const uint32_t &ip_src, const uint32_t &ip_dst, const uint8_t &sport, const uint8_t &dport)
+    void init(const uint32_t &ip_src, const uint32_t &ip_dst, const uint16_t &sport, const uint16_t &dport)
     {
         this->ip_src = ip_src;
         this->ip_dst = ip_dst;
         this->sport = sport;
         this->dport = dport;
     }
+
     void reverse(flowmanage flow)
     {
         this->ip_src = flow.ip_dst;
@@ -42,5 +43,6 @@ public:
                           if(this->dport < flow.dport) return true;
                           else return false;
                   }
+
 };
 #endif // FLOW_H
